remove monitoring for flask
the monitoring code for flask requets was leaking memory
it is removed for now as long as we didn't find out how to prevent this
leak
