+++ zerorobot/robot.py
@@ -16,7 +16,7 @@ from gevent.pywsgi import WSGIServer
 from js9 import j
 from zerorobot import service_collection as scol
 from zerorobot import template_collection as tcol
-from zerorobot.prometheus.flask import monitor
+from zerorobot.prometheus import monitor
 from zerorobot.server.app import app
 from zerorobot.task import PRIORITY_SYSTEM
 from zerorobot import config
